refactor: 레이싱 자동차들 객체
